+++ src/modules/web-app/vitest.config.ts
@@ -1,11 +1,11 @@
 import tsconfigPaths from "vite-tsconfig-paths";
 /// <reference types="vitest" />
-import { configDefaults, defineConfig } from "vitest/config";
+import { defineConfig } from "vitest/config";
 
 export default defineConfig({
   plugins: [tsconfigPaths()],
   test: {
-    environment: "node",
-    exclude: [...configDefaults.exclude, "src/modules/web-app/**/*"],
+    environment: "jsdom",
+    setupFiles: "./setupTests.ts",
   },
 });