+++ tsup.config.ts
@@ -6,5 +6,5 @@ export default defineConfig({
   target: "es2020",
   format: ["esm"],
   outDir: "./build/",
-  onSuccess: "cp -r ./src/modules/WebApp/build/ ./build/web-app/",
+  onSuccess: "cp -r ./src/modules/web-app/build/ ./build/web-app/",
 });