+++ CHANGELOG.md
@@ -6,6 +6,26 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v1.0.0-alpha.9] - 2025-05-26
+
+### Changed
+
+- Updated dependencies
+  - main
+    - Major = vitest, vite-tsconfig-paths
+    - Minor = chalk, zod, @biomejs/biome, tsup, tsx, typescript
+    - Patch = fp-ts, serve-handler, @types/yargs
+  - web-app
+    - Major = vite, vitest
+    - Minor = @testing-library/jest-dom, @testing-library/user-event,
+      @vitejs/plugin-react, eslint-plugin-jest-dom, eslint-plugin-react,
+      postcss, typescript, vite-tsconfig-paths
+    - Patch = clsx, sigma, autoprefixer
+
+### Fixed
+
+- Fixed vulnerabilities that were fixable via (npm audit fix)
+
 ## [v1.0.0-alpha.8] - 2024-08-31
 
 ### Added